+++ rocketgram/api/chat_member.py
@@ -3,7 +3,6 @@
 # Rocketgram is released under the MIT License (see LICENSE).
 
 
-import warnings
 from dataclasses import dataclass
 from datetime import datetime
 from typing import Optional
@@ -54,11 +53,6 @@ class ChatMember:
     can_send_polls: Optional[bool]
     can_send_other_messages: Optional[bool]
     can_add_web_page_previews: Optional[bool]
-
-    @property
-    def can_send_media_messages(self) -> bool:
-        warnings.warn("This field is deprecated.", DeprecationWarning)
-        return self.can_send_other_messages
 
     @classmethod
     def parse(cls, data: dict) -> Optional['ChatMember']: